Refactor editor code to use standardized integer types and improve error handling
- Updated various function signatures to return uint8_t for better error indication.
- Changed integer types in structures and function parameters to use stdint types (e.g., int32_t, uint32_t).
- Replaced all instances of 'int' for boolean flags with 'uint8_t' for clarity.
- Modified the handling of the editor's configuration structure to use a flags union for state management.
- Improved memory management and error handling in functions related to snapshots, rows, and editor operations.
- Ensured consistent use of error handling across the editor functions.
- Updated the terminal handling code to use a more robust signal handling mechanism.
- Removed unnecessary comments and cleaned up code for better readability.

+++ include/buffer.h
@@ -2,16 +2,17 @@
 #define BUFFER_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 struct ABuf {
     char *buf;
-    int len;
+    size_t len;
 };
 
 #define ABUF_INIT \
     { NULL, 0 }
 
-int ab_append(struct ABuf *ab, const char *s, size_t len);
-int ab_free(struct ABuf *ab);
+uint8_t ab_append(struct ABuf *ab, const char *s, size_t len);
+uint8_t ab_free(struct ABuf *ab);
 
 #endif
